fix: enable optionally loading middleware properly
